requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dashscope==1.19.1
distro==1.9.0
fonttools==4.53.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
matplotlib==3.9.0
multidict==6.0.5
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==1.26.4
ollama-python==0.1.2
openai==1.30.1
packaging==24.0
pandas==2.2.2
peewee==3.17.5
pillow==10.3.0
playwright==1.44.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
responses==0.24.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spark-ai-python==0.3.30
tenacity==8.3.0
tencentcloud-sdk-python-common==3.0.1155
tencentcloud-sdk-python-hunyuan==3.0.1155
tencentcloud-sdk-python-lcic==3.0.1155
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
volcengine-python-sdk==1.0.87
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
fastapi>=0.104.0
uvicorn>=0.24.0
celery>=5.3.4
redis>=5.0.1
sqlmodel>=0.0.8
mysqlclient>=2.2.0
